+++ benches/pedersen.rs
@@ -2,11 +2,11 @@
 mod bench_utils;
 
 use ark_std::UniformRand;
-use ark_vrf::{AffinePoint, Input, Secret, pedersen::PedersenSuite};
-use bench_utils::BenchInfo;
-use criterion::{BenchmarkId, Criterion, black_box, criterion_group, criterion_main};
+use ark_vrf::{pedersen::PedersenSuite, AffinePoint, Input, Secret};
+use bench_utils::SuiteExt;
+use criterion::{black_box, criterion_group, criterion_main, BenchmarkId, Criterion};
 
-fn bench_pedersen_prove<S: BenchInfo + PedersenSuite>(c: &mut Criterion) {
+fn bench_pedersen_prove<S: PedersenSuite>(c: &mut Criterion) {
     use ark_vrf::pedersen::Prover;
 
     let secret = Secret::<S>::from_seed([0; 32]);
@@ -19,7 +19,7 @@ fn bench_pedersen_prove<S: BenchInfo + PedersenSuite>(c: &mut Criterion) {
     });
 }
 
-fn bench_pedersen_verify<S: BenchInfo + PedersenSuite>(c: &mut Criterion) {
+fn bench_pedersen_verify<S: PedersenSuite>(c: &mut Criterion) {
     use ark_vrf::pedersen::{Prover, Verifier};
 
     let secret = Secret::<S>::from_seed([0; 32]);
@@ -35,8 +35,8 @@ fn bench_pedersen_verify<S: BenchInfo + PedersenSuite>(c: &mut Criterion) {
 
 const BATCH_SIZES: &[usize] = &[1, 2, 4, 8, 16, 32, 64, 128, 256];
 
-fn bench_pedersen_batch<S: BenchInfo + PedersenSuite>(c: &mut Criterion) {
-    use ark_vrf::pedersen::{BatchItem, BatchVerifier, Prover};
+fn bench_pedersen_batch<S: PedersenSuite>(c: &mut Criterion) {
+    use ark_vrf::pedersen::{BatchVerifier, Prover};
 
     let secret = Secret::<S>::from_seed([0; 32]);
     let max_batch_size = BATCH_SIZES[BATCH_SIZES.len() - 1];
@@ -84,7 +84,7 @@ fn bench_pedersen_batch<S: BenchInfo + PedersenSuite>(c: &mut Criterion) {
     }
 }
 
-fn bench_pedersen_suite<S: BenchInfo + PedersenSuite>(c: &mut Criterion) {
+fn bench_pedersen_suite<S: PedersenSuite>(c: &mut Criterion) {
     bench_pedersen_prove::<S>(c);
     bench_pedersen_verify::<S>(c);
     bench_pedersen_batch::<S>(c);